refactor(threads): useThreadItemEvents 外移谓词纯段至 threadItemEventPredicates.ts
- 2668→2344 行；L93-254 谓词/engine 判定/debug 工具纯段逐字外移
- 三个 hook 契约类型（UseThreadItemEventsOptions 等）留主文件
- CLAUDE_STREAM_DEBUG_FLAG_KEY 随谓词走并回供主文件
- 测试锁面 canProgressEventStartProcessing 主文件 re-export
- baseline 重生成 + 新文件登记豁免
验证：tsc 零 error；useThreadItemEvents 8 文件 79 passed；gate 绿
